multiarch: Implement arch abstraction
- Move preparation step from Makefile to python module
(Different targets need different preparation steps.)
- update build-config for arch abstraction & BuildPreparation step
- remove build-flavor
It's now part of architecture abstraction. Dev packages are
being handled by prepare_common module.

- remove shell script import-local-packages
Ported to python and part of prepare_common module.

- make_version_file
Moved to BuildPreparation package and wrapped in order to
be loadable as a module.

- live_build_config
Moved to BuildPreparation package and wrapped in order to
be loadable as a module.
Passing individual commandline arguments in a list instead of big
string to make argument splitting by improper escaping impossible.